Got rid of player 2 jittering.
